Fix build errors

+++ DevilDaggersWebsite.BlazorWasm/Server/WorldRecords/WorldRecordData.cs
@@ -2,5 +2,5 @@
 
 namespace DevilDaggersWebsite.BlazorWasm.Server.WorldRecords
 {
-	public record WorldRecordData(TimeSpan WorldRecordDuration, int? WorldRecordImprovement);
+	public record WorldRecordData(TimeSpan WorldRecordDuration, double? WorldRecordImprovement);
 }
